Fix compatibility issues with php 8.x

+++ TelegramBot/core/TelegramBot_helper_api.php
@@ -589,7 +589,9 @@ function telegram_add_comment( $p_current_action, $p_message, $p_reply_to_messag
 //                                          'reply_markup' => keyboard_bug_status_change_is( $t_bug_id )
                 ];
             } catch( Mantis\Exceptions\MantisException $t_error ) {
-                $t_file_is_deleted = unlink( $t_file_path );
+                if( isset( $t_file_path ) ) {
+                    $t_file_is_deleted = unlink( $t_file_path );
+                }
 
                 $t_params = $t_error->getParams();
                 if( !empty( $t_params ) ) {
